17 batch write

+++ tutake/api/ts/stk_rewards_ext.py
@@ -1,3 +1,7 @@
+import numpy as np
+
+from tutake.api.ts.date_utils import day_by_day_params
+from tutake.utils import utils
 
 
 def default_cron_express_ext(self) -> str:
@@ -17,7 +21,9 @@ def prepare_ext(self):
     同步历史数据准备工作
     :return:
     """
-    self.delete_all()
+    cnt = self.count()
+    if cnt < 100000:
+        self.delete_all()
 
 
 def query_parameters_ext(self):
@@ -25,7 +31,18 @@ def query_parameters_ext(self):
     同步历史数据调用的参数
     :return: list(dict)
     """
-    return self.api.stock_basic.column_data(['ts_code'])
+    cnt = self.count()
+    if cnt < 1000000:
+        return self.api.stock_basic.column_data(['ts_code'])
+    else:
+        stocks = [s['ts_code'] for s in self.api.stock_basic.column_data(['ts_code'])]
+        stocks = utils.chunks(stocks, 500)
+        dates = day_by_day_params(self, "19911231", date_column="end_date")
+        params = []
+        for date in dates:
+            for stock in stocks:
+                params.append({"end_date": date['end_date'], "ts_code": ','.join(stock)})
+        return params
 
 
 def param_loop_process_ext(self, **params):
